+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add support for shared memory limit metrics from nchan >= 0.1.10
+
 ## 0.1.0
 
 * Initial release.